+++ app/controllers/application_controller.rb
@@ -1,4 +1,23 @@
 class ApplicationController < ActionController::Base
+  helper_method :current_user, :logged_in?
   add_flash_types :success, :info, :warning, :danger
-  include SessionsHelper
+
+  private
+
+    def log_in(user)
+      session[:user_id] = user.id
+    end
+
+    def current_user
+      @current_user ||= User.find_by(id: session[:user_id])
+    end
+
+    def logged_in?
+      current_user.present?
+    end
+
+    def log_out
+      session.delete(:user_id)
+      @current_user = nil
+    end
 end
